refactor(#602): Random task executive
- Updates from rcppsw#183, and also a few warning fixes from running the intel
  compiler.

+++ src/support/depth1/static_cache_creator.cpp
@@ -71,9 +71,9 @@ ds::cache_vector static_cache_creator::create_all(
 
     if (static_cast<uint>(std::distance(it, c_alloc_blocks.end())) <
         base_cache::kMinBlocks) {
-      ER_WARN("Not enough blocks provided to construct cache@%s: %zu < %zu",
+      ER_WARN("Not enough blocks provided to construct cache@%s: %u < %zu",
               center.to_str().c_str(),
-              std::distance(it, c_alloc_blocks.end()),
+              static_cast<uint>(std::distance(it, c_alloc_blocks.end())),
               base_cache::kMinBlocks);
       continue;
     }
